Changed w2v script

+++ narrowing_ai_research/estimators/train_word2vec.py
@@ -12,15 +12,25 @@ project_dir = narrowing_ai_research.project_dir
 with open(f"{project_dir}/model_config.yaml", "r") as infile:
     min_count = yaml.safe_load(infile)["min_count"]
 
-logging.info("loading and processing data")
-with open(f"{project_dir}/data/interim/arxiv_tokenised.json", "r") as infile:
-    arxiv_tokenised = json.load(infile)
 
-tok = list(arxiv_tokenised.values())
+def train_word2vec():
 
-logging.info("Training model")
-ft = FastText(tok, min_count=min_count, word_ngrams=0)
+    if os.path.exists(f"{project_dir}/models/word2vec.p") is True:
+        logging.info("Model already trained")
+    else:
+        logging.info("loading and processing data")
+        with open(f"{project_dir}/data/interim/arxiv_tokenised.json", "r") as infile:
+            arxiv_tokenised = json.load(infile)
 
-# Save model
-with open(f"{project_dir}/models/word2vec.p", "wb") as outfile:
-    pickle.dump(ft, outfile)
+        tok = list(arxiv_tokenised.values())
+
+        logging.info("Training model")
+        ft = FastText(tok, min_count=min_count, word_ngrams=0)
+
+        # Save model
+        with open(f"{project_dir}/models/word2vec.p", "wb") as outfile:
+            pickle.dump(ft, outfile)
+
+
+if __name__ == "__main__":
+    train_word2vec()
